# lang-portal/backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.AddSessionActivityRequest:
    properties:
      activity_id:
        minimum: 1
        type: integer
      answer:
        type: string
      session_id:
        minimum: 1
        type: integer
    required:
    - activity_id
    - answer
    - session_id
    type: object
...
  handlers.EvaluateSessionActivityRequest:
    properties:
      result:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
    required:
    - result
    type: object
...
  models.Group:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.SessionActivity:
    properties:
      activity_id:
        type: integer
      answer:
        type: string
      created_at:
        type: string
      id:
        type: integer
      result:
        type: string
      score:
        type: integer
      session_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@languageportal.com
    name: Language Portal Support
  description: A backend API for managing language learning resources
  title: Language Portal API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Retrieve a paginated list of groups with optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Groups retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group with a name and optional description
      parameters:
      - description: Group details
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "201":
          description: Group created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or validation error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Delete an existing group by its unique identifier
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a group
      tags:
      - groups
    get:
      description: Retrieve a group's details using its unique identifier
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update an existing group's details by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated group details
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: Group updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a group
      tags:
      - groups
  /session-activities/{id}:
    delete:
      description: Delete a specific session activity
      parameters:
      - description: Session Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Session activity deleted successfully
        "400":
          description: Invalid session activity ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete session activity
      tags:
      - session-activities
    put:
      consumes:
      - application/json
      description: Evaluate and score a session activity
      parameters:
      - description: Session Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Evaluation details
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/handlers.EvaluateSessionActivityRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Session activity evaluated successfully
        "400":
          description: Invalid request payload or session activity ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Evaluate session activity
      tags:
      - session-activities
  /sessions/{session_id}/activities:
    get:
      description: Retrieve all activities for a specific session
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session activities retrieved successfully
          schema:
            items:
              $ref: '#/definitions/models.SessionActivity'
            type: array
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List session activities
      tags:
      - session-activities
    post:
      consumes:
      - application/json
      description: Add a new activity to an existing session
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: integer
      - description: Session activity details
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/handlers.AddSessionActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session activity added successfully
          schema:
            $ref: '#/definitions/models.SessionActivity'
        "400":
          description: Invalid request payload or session/activity ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add session activity
      tags:
      - session-activities
...
schemes:
- http
swagger: "2.0"